features 3 and 4 are added. And updated logic + refactored code for all
features from 1 to 4

+++ controller/post/postController.js
@@ -7,6 +7,9 @@ const {
   generateTimedToken,
 } = require("../../utils/tokens");
 require("dotenv").config();
+const Notification = require("../../models/notificationModel");
+const { getVoteStatusAndSubredditDetails } = require("../../utils/posts");
+
 
 // Function to retrieve all comments for a post.
 /**
@@ -22,7 +25,7 @@ async function getPostComments(req, res) {
   try {
     const postId = decodeURIComponent(req.params.postId);
 
-    const post = await Post.findById(postId);
+    const post = await Post.findById(postId).populate("originalPostId");
 
     if (!post) {
       return res
@@ -62,7 +65,7 @@ async function createComments(req, res) {
   try {
     const { postId, content } = req.body;
 
-    const post = await Post.findById(postId);
+    const post = await Post.findById(postId).populate("originalPostId");
 
     if (!post) {
       return res
@@ -70,26 +73,48 @@ async function createComments(req, res) {
         .json({ success: false, message: "Post not found." });
     }
 
-      // Check if the post is locked
-      if (post.isLocked) {
-        return res.status(403).json({ success: false, message: "Post is locked. Cannot add a comment." });
-      }
+    // Check if the post is locked
+    if (post.isLocked) {
+      return res
+        .status(403)
+        .json({
+          success: false,
+          message: "Post is locked. Cannot add a comment.",
+        });
+    }
 
-      const comment = new Comment({
-        content,
-        authorName: user.username,
-        createdAt: new Date(),
-        upvotes: 0,
-        downvotes: 0,
-        linkedPost: postId,
-        linkedSubreddit: post.linkedSubreddit,
-        awards: 0
-      });
-  
-      await comment.save();
-  
-      return res.status(201).json({ success: true, comment });
+      // check if user is authorized to comment
+      // if (post.linkedSubreddit.privacyMode === "private") {
+      //   const isMember = post.linkedSubreddit.members.includes(user._id);
+      //   if (!isMember) {
+      //     return res.status(403).json({ success: false, message: "You are not authorized to comment on this post." });
+      //   }
+      // }
 
+    const comment = new Comment({
+      content,
+      authorName: user.username,
+      createdAt: new Date(),
+      upvotes: 0,
+      downvotes: 0,
+      linkedPost: postId,
+      linkedSubreddit: post.linkedSubreddit,
+      awards: 0,
+    });
+
+    await comment.save();
+    post.comments.push(comment._id);
+    await post.save();
+
+    // Create a notification for the post author
+    const notification = new Notification({
+      title: "New Comment",
+      message: `${user.username} commented on your post "${post.title}".`,
+      recipient: post.authorName, // Assuming `author` is the username of the post author
+    }); 
+  
+    await notification.save();
+    return res.status(201).json({ success: true, comment });
   } catch (err) {
     console.log(err);
     return res
@@ -122,6 +147,11 @@ async function updatePostComments(req, res) {
 
     if (!comment) {
       return res.status(404).json({ success: false,  message: "comment not found" });
+    }
+
+    // check if user is authorized to edit comment
+    if (comment.authorName !== user.username) {
+      return res.status(403).json({ success: false, message: "You are not authorized to edit this comment." });
     }
 
     comment.content = content;
@@ -162,7 +192,21 @@ async function deleteComments(req, res) {
     if (!comment) {
       return res.status(404).json({ success: false,  message: "comment not found" });
     }
+
+    // Check if the user is the author of the comment
+    const isAuthor = comment.authorName === user.username;
+
+    if (!isAuthor) {
+      return res.status(403).json({ success: false, message: "You are not authorized to delete this comment." });
+    }
+
     await Comment.deleteOne({ _id: commentId });
+
+    await Post.updateOne(
+      { _id: comment.linkedPost },
+      { $pull: { comments: commentId } }
+    );
+
     return res.status(200).json({ success: true, message: "comment deleted successfully" });
 
   } catch (err) {
@@ -173,6 +217,14 @@ async function deleteComments(req, res) {
   }
 }
 
+// Function to delete a post
+/**
+ * Deletes a post.
+ * @async
+ * @param {object} req - Express request object
+ * @param {object} res - Express response object
+ * @returns {object} - Express response object
+ */
 
 async function deletePost (req, res) {
   const token = req.headers.authorization.split(" ")[1];
@@ -186,10 +238,16 @@ async function deletePost (req, res) {
   }
   try {
     const postId = req.params.postId;
-    const post = await Post.findById(postId);
+    const post = await Post.findById(postId).populate("originalPostId");
     if (!post) {
       return res.status(404).json({ success: false, message: "Post not found." });
     } 
+    // Check if the user is the author of the post
+    const isAuthor = post.authorName === user.username;
+    if (!isAuthor) {
+        return res.status(403).json({ success: false, message: "You are not authorized to delete this post." });
+    }
+
     await post.deleteOne();
     return res.status(200).json({ success: true, message: "Post deleted successfully." });
   }
@@ -198,6 +256,15 @@ async function deletePost (req, res) {
     return res.status(500).json({ success: false, message: "Internal server error." });
   }
 }
+
+// Function to edit post content
+/**
+ * Edits post content.
+ * @async
+ * @param {object} req - Express request object
+ * @param {object} res - Express response object
+ * @returns {object} - Express response object
+ */
 
 async function editPostContent(req, res) {
   const token = req.headers.authorization.split(" ")[1];
@@ -211,19 +278,33 @@ async function editPostContent(req, res) {
   }
   try {
     const { postId, content } = req.body;
-    const post = await Post.findById(postId);
+    const post = await Post.findById(postId).populate("originalPostId");
     if (!post) {
       return res.status(404).json({ success: false, message: "Post not found." });
     }
+    // check if user is authorized to edit post
+    if (post.authorName !== user.username) {
+      return res.status(403).json({ success: false, message: "You are not authorized to edit this post." });
+    }
+
     post.content = content;
     await post.save();
-    return res.status(200).json({ success: true, post });
 
+    return res.status(200).json({ success: true, post });
   } catch (err) {
     console.log(err);
     return res.status(500).json({ success: false, message: "Internal server error." });
   }
 }
+
+// Function to mark a post as NSFW
+/**
+ * Marks a post as NSFW (Not Safe For Work).
+ * @async
+ * @param {object} req - Express request object
+ * @param {object} res - Express response object
+ * @returns {object} - Express response object
+ */
 
 // NSFW = Not Safe For Work
 async function markPostNSFW(req, res) {
@@ -238,10 +319,11 @@ async function markPostNSFW(req, res) {
   }
   try {
     const { postId } = req.body;
-    const post = await Post.findById(postId);
+    const post = await Post.findById(postId).populate("originalPostId");
     if (!post) {
       return res.status(404).json({ success: false, message: "Post not found." });
     }
+    
     post.isNSFW = true;
     await post.save();
     return res.status(200).json({ success: true, post });
@@ -251,6 +333,15 @@ async function markPostNSFW(req, res) {
     return res.status(500).json({ success: false, message: "Internal server error." });
   }
 }
+
+// Function to unmark a post as NSFW
+/**
+ * Unmarks a post as NSFW (Not Safe For Work).
+ * @async
+ * @param {object} req - Express request object
+ * @param {object} res - Express response object
+ * @returns {object} - Express response object
+ */
 
 async function unmarkPostNSFW(req, res) {
   const token = req.headers.authorization.split(" ")[1];
@@ -264,7 +355,7 @@ async function unmarkPostNSFW(req, res) {
   }
   try {
     const { postId } = req.body;
-    const post = await Post.findById(postId);
+    const post = await Post.findById(postId).populate("originalPostId");
     if (!post) {
       return res.status(404).json({ success: false, message: "Post not found." });
     }
